Fixed issue with invisible buttons
Created a variable "change_click_through" that when in true, changes the
click_through of a button to true when it is disabled and to false when
it is enabled.
This solves a problem that made that some buttons could not get clicked
because another invisible button was on top of the first with the
click_through on false.